mcp/requirements.txt
azure-functions
azure-identity>=1.17.0
azure-search-documents>=11.6.0
